Move accessor container indexing.hpp to definitions directory
