+++ core/include/detray/core/detector.hpp
@@ -14,6 +14,7 @@
 #include "detray/core/mask_store.hpp"
 #include "detray/core/surfaces_finder.hpp"
 #include "detray/core/transform_store.hpp"
+#include "detray/definitions/detail/accessor.hpp"
 #include "detray/definitions/qualifiers.hpp"
 #include "detray/geometry/surface.hpp"
 #include "detray/geometry/volume.hpp"
@@ -22,7 +23,6 @@
 #include "detray/grids/populator.hpp"
 #include "detray/grids/serializer2.hpp"
 #include "detray/intersection/intersection.hpp"
-#include "detray/utils/detail/accessor.hpp"
 
 namespace detray {
