+++ core/include/detray/core/transform_store.hpp
@@ -8,9 +8,9 @@
 
 #include <vecmem/memory/memory_resource.hpp>
 
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
 #include "detray/utils/enumerate.hpp"
-#include "detray/utils/indexing.hpp"
 
 namespace detray {
