+++ core/include/detray/definitions/detail/accessor.hpp
@@ -20,8 +20,8 @@
 #include <type_traits>
 #include <utility>
 
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
-#include "detray/utils/indexing.hpp"
 
 namespace detray {
