+++ core/include/detray/geometry/surface.hpp
@@ -8,9 +8,9 @@
 
 #include <memory>
 
+#include "detray/definitions/detail/accessor.hpp"
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
-#include "detray/utils/detail/accessor.hpp"
-#include "detray/utils/indexing.hpp"
 
 namespace detray {
