+++ core/include/detray/grids/populator.hpp
@@ -8,8 +8,8 @@
 #pragma once
 
 // Project include(s).
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
-#include "detray/utils/indexing.hpp"
 #include "detray/utils/invalid_values.hpp"
 
 // VecMem include(s).
