+++ core/include/detray/intersection/concentric_cylinder_intersector.hpp
@@ -10,11 +10,11 @@
 #include <cmath>
 #include <type_traits>
 
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
 #include "detray/intersection/intersection.hpp"
 #include "detray/intersection/quadratic_equation.hpp"
 #include "detray/propagator/track.hpp"
-#include "detray/utils/indexing.hpp"
 
 namespace detray {
