+++ core/include/detray/intersection/intersection.hpp
@@ -9,8 +9,8 @@
 #include <climits>
 #include <sstream>
 
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
-#include "detray/utils/indexing.hpp"
 
 namespace detray {
