+++ core/include/detray/intersection/intersection_kernel.hpp
@@ -10,12 +10,12 @@
 #include <tuple>
 #include <utility>
 
+#include "detray/definitions/detail/accessor.hpp"
+#include "detray/definitions/indexing.hpp"
 #include "detray/definitions/qualifiers.hpp"
 #include "detray/intersection/intersection.hpp"
 #include "detray/propagator/track.hpp"
-#include "detray/utils/detail/accessor.hpp"
 #include "detray/utils/enumerate.hpp"
-#include "detray/utils/indexing.hpp"
 
 namespace detray {
