+++ core/include/detray/masks/single3.hpp
@@ -11,10 +11,10 @@
 #include <sstream>
 #include <string>
 
+#include "detray/definitions/detail/accessor.hpp"
 #include "detray/definitions/qualifiers.hpp"
 #include "detray/intersection/intersection.hpp"
 #include "detray/intersection/planar_intersector.hpp"
-#include "detray/utils/detail/accessor.hpp"
 
 namespace detray {
 /** This is a simple mask for single parameter bound mask
