+++ tests/benchmarks/core/benchmark_grids.cpp
@@ -12,11 +12,11 @@
 #include "tests/common/test_defs.hpp"
 
 // detray core
+#include "detray/definitions/indexing.hpp"
 #include "detray/grids/axis.hpp"
 #include "detray/grids/grid2.hpp"
 #include "detray/grids/populator.hpp"
 #include "detray/grids/serializer2.hpp"
-#include "detray/utils/indexing.hpp"
 
 using namespace detray;
 using namespace __plugin;
