+++ tests/unit_tests/core/grids_axis.cpp
@@ -15,8 +15,8 @@
 #include "tests/common/test_defs.hpp"
 
 // detray core
+#include "detray/definitions/indexing.hpp"
 #include "detray/grids/axis.hpp"
-#include "detray/utils/indexing.hpp"
 
 using namespace detray;
