+++ tests/unit_tests/core/grids_grid2.cpp
@@ -15,11 +15,11 @@
 
 #include <climits>
 
+#include "detray/definitions/indexing.hpp"
 #include "detray/grids/axis.hpp"
 #include "detray/grids/grid2.hpp"
 #include "detray/grids/populator.hpp"
 #include "detray/grids/serializer2.hpp"
-#include "detray/utils/indexing.hpp"
 
 using namespace detray;
