+++ tests/unit_tests/core/grids_populator.cpp
@@ -13,8 +13,8 @@
 #include "tests/common/test_defs.hpp"
 
 // detray core
+#include "detray/definitions/indexing.hpp"
 #include "detray/grids/populator.hpp"
-#include "detray/utils/indexing.hpp"
 
 using namespace detray;
