+++ tests/unit_tests/core/grids_serializer.cpp
@@ -15,9 +15,9 @@
 #include "tests/common/test_defs.hpp"
 
 // detray core
+#include "detray/definitions/indexing.hpp"
 #include "detray/grids/axis.hpp"
 #include "detray/grids/serializer2.hpp"
-#include "detray/utils/indexing.hpp"
 
 using namespace detray;
