+++ tests/unit_tests/cuda/grids_grid2_cuda_kernel.hpp
@@ -9,12 +9,12 @@
 #include "tests/common/test_defs.hpp"
 
 // detray core
+#include "detray/definitions/indexing.hpp"
 #include "detray/grids/axis.hpp"
 #include "detray/grids/grid2.hpp"
 #include "detray/grids/populator.hpp"
 #include "detray/grids/serializer2.hpp"
 #include "detray/tools/grid_array_helper.hpp"
-#include "detray/utils/indexing.hpp"
 
 #pragma once
